test(cli): improve contract test cohesion
- split test profile init process-contract tests into dedicated class

- extract CLI process runner and JSON contract assertions into shared helpers

- keep init/status contract tests focused on their command responsibilities
